oct 9th 2017 thinkpad reinstall

+++ kernel_recompile.sh
@@ -1,6 +1,8 @@
 #!/bin/sh
 
 #mount | grep '/boot' || { echo "mount /boot first. Exiting." && exit 1 ; }
+test -s /boot/grub/grub.cfg || { echo "/boot/grub/grub.cfg not found. Exiting." ; exit 1 ; }
+
 ls /boot/vmlinuz || { echo "mount /boot first. Exiting." && exit 1 ; }
 
 test -e /usr/src/linux/.config || ln -s /home/cfg/sysskel/usr/src/linux_configs/.config /usr/src/linux/.config
@@ -10,3 +12,5 @@ echo "kernel compile and install completed OK"
 
 qpkg zfs && USE="${USE} -kernel-builtin" emerge spl zfs zfs-kmod || { echo "xfs is not installed, skipping \"emerge zfs\"" ; }
 qpkg zfs && genkernel initramfs --no-clean --no-mountboot --zfs --symlink
+grub-mkconfig -o /boot/grub/grub.cfg
+
